requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.19.4
matplotlib==3.5.1
